Create separate Rhel ecosystem
Rhel ecosystem includes rhel specific package, dist scanners and custom
coalescer that handles CPE based repositories. Most of coalescer code is
similar to the linux one. The rhel coalescer was created to not
introduce rhel specific cases into generic linux coalescer.
