+++ internal/indexer/linux/coalescer_test.go
@@ -91,85 +91,3 @@ func Test_Coalescer(t *testing.T) {
 		}
 	}
 }
-
-func Test_Coalescer_rhel(t *testing.T) {
-	ctx, done := context.WithCancel(context.Background())
-	defer done()
-	ctx = log.TestLogger(ctx, t)
-	coalescer := &Coalescer{
-		ir: &claircore.IndexReport{
-			Environments:  map[string][]*claircore.Environment{},
-			Packages:      map[string]*claircore.Package{},
-			Distributions: map[string]*claircore.Distribution{},
-			Repositories:  map[string]*claircore.Repository{},
-		},
-	}
-	repo1 := &claircore.Repository{
-		ID:   "1",
-		Name: "cpe:/o:redhat:enterprise_linux:8::baseos",
-		Key:  "rhel-cpe-repo",
-	}
-	repo2 := &claircore.Repository{
-		ID:   "2",
-		Name: "cpe:/o:redhat:enterprise_linux:8::appstream",
-		Key:  "rhel-cpe-repo",
-	}
-	repo3 := &claircore.Repository{
-		ID:   "3",
-		Name: "cpe:/o:redhat:enterprise_linux:8::appstream",
-		Key:  "rhel-cpe-repo",
-	}
-
-	pkgs := test.GenUniquePackages(5)
-	dists := test.GenUniqueDistributions(3) // we will discard dist 0 due to zero value ambiguity
-	layerArtifacts := []*indexer.LayerArtifacts{
-		{
-			Hash:  test.RandomSHA256Digest(t),
-			Pkgs:  pkgs[0:1],
-			Dist:  nil,
-			Repos: nil,
-		},
-		{
-			Hash:  test.RandomSHA256Digest(t),
-			Pkgs:  pkgs[1:2],
-			Dist:  nil,
-			Repos: []*claircore.Repository{repo1, repo2},
-		},
-		{
-			Hash:  test.RandomSHA256Digest(t),
-			Pkgs:  pkgs[2:3],
-			Dist:  dists[1:2],
-			Repos: []*claircore.Repository{repo3},
-		},
-		{
-			Hash:  test.RandomSHA256Digest(t),
-			Pkgs:  pkgs[3:4],
-			Dist:  nil,
-			Repos: nil,
-		},
-		{
-			Hash:  test.RandomSHA256Digest(t),
-			Pkgs:  pkgs[4:5],
-			Dist:  dists[2:],
-			Repos: nil,
-		},
-	}
-	ir, err := coalescer.Coalesce(ctx, layerArtifacts)
-	if err != nil {
-		t.Fatalf("received error from coalesce method: %v", err)
-	}
-	// we expect packages 1-2 to be associated with repos 1 and 2
-	for i := 0; i < 2; i++ {
-		environment := ir.Environments[strconv.Itoa(i)][0]
-		if len(environment.RepositoryIDs) != 2 || environment.RepositoryIDs[0] != "1" || environment.RepositoryIDs[1] != "2" {
-			t.Fatalf("expected repository ids [1, 2] but got %s", environment.RepositoryIDs)
-		}
-	}
-	// and packages 3-5 to be associated with repo 3
-	for i := 2; i < 5; i++ {
-		environment := ir.Environments[strconv.Itoa(i)][0]
-		if len(environment.RepositoryIDs) != 1 || environment.RepositoryIDs[0] != "3" {
-			t.Fatalf("expected repository ids [3] but got %s", environment.RepositoryIDs)
-		}
-	}
-}
