+++ rhel/ecosystem.go
@@ -1,35 +1,34 @@
-package rpm
+package rhel
 
 import (
 	"context"
 
-	"github.com/quay/claircore/aws"
 	"github.com/quay/claircore/internal/indexer"
-	"github.com/quay/claircore/internal/indexer/linux"
-	"github.com/quay/claircore/oracle"
-	"github.com/quay/claircore/photon"
-	"github.com/quay/claircore/suse"
+	"github.com/quay/claircore/rpm"
 )
 
-// NewEcosystem provides the set of scanners and coalescers for the rpm ecosystem
+// NewEcosystem provides the set of scanners and coalescers for the rhel ecosystem
 func NewEcosystem(ctx context.Context) *indexer.Ecosystem {
 	return &indexer.Ecosystem{
 		PackageScanners: func(ctx context.Context) ([]indexer.PackageScanner, error) {
-			return []indexer.PackageScanner{&Scanner{}}, nil
+			return []indexer.PackageScanner{
+				&rpm.Scanner{},
+			}, nil
 		},
 		DistributionScanners: func(ctx context.Context) ([]indexer.DistributionScanner, error) {
 			return []indexer.DistributionScanner{
-				&aws.DistributionScanner{},
-				&oracle.DistributionScanner{},
-				&suse.DistributionScanner{},
-				&photon.DistributionScanner{},
+				&DistributionScanner{},
 			}, nil
 		},
 		RepositoryScanners: func(ctx context.Context) ([]indexer.RepositoryScanner, error) {
-			return []indexer.RepositoryScanner{}, nil
+			return []indexer.RepositoryScanner{
+				&RepositoryScanner{
+					CpeFetcher: &ContainerAPICpeFetcher{},
+				},
+			}, nil
 		},
 		Coalescer: func(ctx context.Context) (indexer.Coalescer, error) {
-			return linux.NewCoalescer(), nil
+			return NewCoalescer(), nil
 		},
 	}
 }
